*** empty log message ***

+++ src/chm.h
@@ -9,7 +9,7 @@ typedef struct __chm_config_data_t chm_config_data_t;
 chm_config_data_t *chm_config_new();
 void chm_config_set_hashfuncs(cmph_config_t *mph, CMPH_HASH *hashfuncs);
 void chm_config_destroy(cmph_config_t *mph);
-cmph_t *chm_new(cmph_config_t *mph, float c);
+cmph_t *chm_new(cmph_config_t *mph, double c);
 
 void chm_load(FILE *f, cmph_t *mphf);
 int chm_dump(cmph_t *mphf, FILE *f);
